packages/dcor-shared/dcor_shared-0.8.0.tar.gz/dcor_shared-0.8.0/tests/test_testing.py
import pathlib
import pytest

import ckan.tests.factories as factories

from dcor_shared.testing import make_dataset

data_path = pathlib.Path(__file__).parent / "data"


@pytest.mark.ckan_config('ckan.plugins', 'dcor_schemas')
@pytest.mark.usefixtures('clean_db', 'with_request_context')
def test_make_dataset():
    user = factories.User()
    owner_org = factories.Organization(users=[{
        'name': user['id'],
        'capacity': 'admin'
    }])
    # Note: `call_action` bypasses authorization!
    create_context = {'ignore_auth': False,
                      'user': user['name'], 'api_version': 3}
    ds_dict = make_dataset(create_context, owner_org,
                           activate=False)
    assert ds_dict["state"] == "draft"


@pytest.mark.ckan_config('ckan.plugins', 'dcor_schemas')
@pytest.mark.usefixtures('clean_db', 'with_request_context')
def test_make_dataset_no_user_specified():
    ds_dict = make_dataset(activate=False)
    assert ds_dict["state"] == "draft"


@pytest.mark.ckan_config('ckan.plugins', 'dcor_schemas')
@pytest.mark.usefixtures('clean_db', 'with_request_context')
def test_make_dataset_with_resource(create_with_upload):
    # upload a *valid* [sic] .rtdc File (this is the control)
    ds_dict, res_dict = make_dataset(
        create_with_upload=create_with_upload,
        resource_path=data_path / "calibration_beads_47.rtdc",
        activate=True)
    assert len(ds_dict["resources"]) == 1
    assert "id" in res_dict
    assert res_dict["package_id"] == ds_dict["id"]
